Split NSGA3 code
