+++ optirustic/src/algorithms/nsga3/test_data/Normalise_normalised_objectives.csv
@@ -1,0 +1,33 @@
+,Objective 0,Objective 1,Objective 2
+Individual 0,0.084517,0.303721,0.661099
+Individual 1,0.511117,0.245795,0.2791
+Individual 2,0.244495,0.119069,0.597467
+Individual 3,0.529209,0.277032,0.299043
+Individual 4,0.103514,0.485039,0.556837
+Individual 5,0.233327,0.154105,0.61871
+Individual 6,0.315263,0.818765,0.007977
+Individual 7,0.007493,0.0,0.207809
+Individual 8,0.064114,0.014397,0.209565
+Individual 9,0.04802,0.08851,0.865264
+Individual 10,0.2671,0.719904,0.0
+Individual 11,0.141006,0.212055,0.723938
+Individual 12,0.014512,0.001417,0.257062
+Individual 13,0.266655,0.718828,0.000913
+Individual 14,0.221828,0.05521,0.591239
+Individual 15,0.206265,0.129243,0.676056
+Individual 16,0.233327,0.154105,0.61871
+Individual 17,0.14101,0.21205,0.723938
+Individual 18,0.267092,0.719916,0.0
+Individual 19,0.510134,0.245328,0.28013
+Individual 20,0.178727,0.044301,0.488425
+Individual 21,0.170309,0.144872,0.42946
+Individual 22,0.659897,0.236375,0.103729
+Individual 23,0.103007,0.483097,0.554434
+Individual 24,0.012954,0.040233,0.751057
+Individual 25,0.014008,0.001285,0.257534
+Individual 26,0.197828,0.34759,0.210451
+Individual 27,0.109866,0.375851,0.634508
+Individual 28,0.0,0.005525,0.994475
+Individual 29,0.150934,0.830999,0.018068
+Individual 30,0.309578,0.325566,0.249266
+Individual 31,0.139888,0.089296,0.457349
