+++ optirustic/src/algorithms/nsga3/test_data/Normalise_objectives.csv
@@ -1,0 +1,33 @@
+,Objective 0,Objective 1,Objective 2
+Individual 0,27.884165,51.326162,226.643641
+Individual 1,136.109674,41.819759,105.633536
+Individual 2,68.469587,21.022148,206.486078
+Individual 3,140.699615,46.946159,111.950948
+Individual 4,32.703566,81.083085,193.615367
+Individual 5,65.636198,26.772029,213.21536
+Individual 6,86.422955,135.852146,19.746699
+Individual 7,8.343896,1.481285,83.04971
+Individual 8,22.708077,3.843962,83.606085
+Individual 9,18.62526,16.007101,291.319088
+Individual 10,74.204324,119.627674,17.219842
+Individual 11,42.215237,36.28256,246.549774
+Individual 12,10.124442,1.713838,98.652123
+Individual 13,74.09151,119.451112,17.509219
+Individual 14,62.718992,10.542082,204.513056
+Individual 15,58.770891,22.691955,231.381681
+Individual 16,65.636198,26.772029,213.21536
+Individual 17,42.216061,36.281736,246.549774
+Individual 18,74.202289,119.62971,17.219842
+Individual 19,135.860262,41.743127,105.959581
+Individual 20,51.78473,8.751678,171.943608
+Individual 21,49.649132,25.256854,153.264594
+Individual 22,173.85407,40.273678,50.079148
+Individual 23,32.574991,80.764304,192.854162
+Individual 24,9.729123,8.08408,255.140458
+Individual 25,9.99653,1.692186,98.801688
+Individual 26,56.630442,58.52572,83.886697
+Individual 27,34.315195,63.163708,218.219944
+Individual 28,6.442876,2.388091,332.250554
+Individual 29,44.733712,137.859885,22.943305
+Individual 30,84.980677,54.911215,96.182449
+Individual 31,41.931455,16.136064,162.099259
